Add C++ validation harness (#12)
* Fix default protoc in cc_proto_library

The old default used a relative reference. Since this is a fork from the
original protobuf repo location, use an absolute reference instead.

* bazel: test executor depends on harness as data

The harness is an executable, not a library that is imported, so treat
it as a generated data item.

* Build C++ test protos

* Support lax test harnesses

This will be used to prevent the tests from all failing once the C++
validation harness is introduced.

* Add C++ test harness

* Bring naming in line with Google/Envoy style

* Allocate message for validation inside lambda

Instead of allocating the memory separately and then using it in the
lambda, allocate the destination proto message inside the lambda because
it's cleaner and doesn't mix scopes.