Shuffle go-harness binary around to satisfy tests
The new version of go_rules seems to have changed the location of the
output binary causing the tests that rely on the location of said binary
to fail. As a quick and dirty fix we use a genrule to move the file into
the location that the tests except it to be in. The real solution is to
pass the location of the binary to the executor as an argument (using
$(location))