standardize range notation in README
unsure if there is any specification or purpose to the various ways the README describes a range. But seeing all of :
` range [30, 40)`
` range [30, 40]`
` range (x, y) `
etc, leads me to think it was a typo instead of syntax. Apologies if this is a naive mistake on my end!

this change lands on `[, ]` as the common range notation.