README: fix example.proto and protoc includes
The extra -I are needed to have the

    import "validate/validate.proto";

work -- and it needs to be imported like this because that's how the
filedescriptor is registered, see

https://github.com/lyft/protoc-gen-validate/blob/e60e7f91da4/validate/validate.pb.go#L3049

Also the proto file needs package, otherwise, you'll see

    protoc-gen-go: error:can't find object with type ..Person

And those character classes in the regex don't seem to be supported?
Using them, I ended up seeing

    example.proto:11:39: Invalid escape sequence in string literal.
    example.proto:11:41: Invalid escape sequence in string literal.
    example.proto:11:49: Invalid escape sequence in string literal.
    example.proto:11:51: Invalid escape sequence in string literal.

This should fix #56.

Signed-off-by: Stephan Renatus <srenatus@chef.io>