Fix Bazel build by generating proto code automatically (#96)
The test code depends on auto-generated Go code that, if missing, will cause compilation errors. This can be generated via make, but doing so

1. is not hermetic
2. requires proto path variables to be set

This PR updates the Bazel rules and targets so that bazel run //tests/harness/executor works again and runs the full test suite. The changes are limited to BUILD files (there are no code changes) so it should only affect Bazel-based builds.