bazel: Rewrite java rule to generate code for transitive protos
Fixes #135

The implementation now:
* Generates all java validation code directly in a rule using
java_common.compile.
* No longer generates java_proto code
* Uses an aspect to generate validation code for transitive
proto dependencies.
This is not strictly necessary (at compile time), but by doing
this the rule makes sure that at runtime the correct validators
are on the classpath.

Validate java targets should now always be of the form

    pgv_java_proto_library(
        name = "a_java_validate",
        deps = [":a_proto"]
        java_deps = [":a_java_proto"]
    )

Note that this change is limited to an implementation for the Java plugin,
It should be possible to abstract most of the logic and reuse it for the
other plugins.