Changed UUID for C++ and Go
I tried to test things but could not get it to work. Not with a local Bazel and not with the Docker image and I am giving up now. Locally the problem was with the `maven_jar` instruction in the `repositories.bzl` file and the Docker image is missing executables, namely gogofast. I also have to say that working on the C++ and Go parts of the project is terrible (sorry for the harsh choice of words but it is the truth). C++ with Go templates? No syntax highlighting? No normal tooling works? … 😟

Signed-off-by: Fleshgrinder <fleshgrinder@users.noreply.github.com>