Create a separate cc_library for proto source (#393)
Since the generated PGV code uses static registration, it needs to be
compiled in a cc_library rule with alwayslink set to 1. Before this
change, that library also included protobuf code, which means that all
generated code for any message type was being included in any consuming
binaries. This patch moves the generated .cc and .h files into a
separate library that the linker can prune symbols from.

Signed-off-by: Alex Konradi <akonradi@google.com>