go.mod: bump protoc-gen-star to v0.5.3
By bumping the version we gain the ability to add the supported features
to the code generation response which helps by not erroring out when it
gets called with the newly released protobuf 3.15. As the experimental
flag for optionals was made mandatory if an `optional` clause exist in
the source proto files.

Signed-off-by: Gorka Lerchundi Osa <glertxundi@gmail.com>