Fix ignore_empty test cases.
Prior to this change, ignore_empty test cases where testing an negated
range (<=128, >=256), and since the value 0 falls in that range, the
tests would pass regardless of ignore_empty.  To make the tests pass
for the right reason (ignore_empty is given), the range is changed to
[128,256] so 0 falls outside the range.

The tests for Float already had the correct range. This commit updates
all the other messages to be consistent with Float.

Signed-off-by: Nadav Samet <thesamet@gmail.com>