make templates self-contained and pluginable
This change introduces templates as plugins, meaning consumers of PGV
can now provide a go-plugin to serve as a template instead of relying in
the built-in ones. Note the ones that were built in were slightly
refactored to compose a `TemplatePlugin` type but functionality were
left untouched and are embedded the same way to reduce the blast radius
of this change.

Signed-off-by: Martin Conte Mac Donell <reflejo@gmail.com>