C++: Fix polymorphic lookup and add tests
The previous iteration does not make ASAN happy because of the implicit
static cast between Message and concrete type. This PR reverts that
change, adds a new abstract validation registry, and explicitly performs
a dynamic cast when asked to do so. It also checks to make sure a
validator exists for the passed in message.

Also add a .clang-format so at least manual formatting can be done.

Signed-off-by: Matt Klein <mklein@lyft.com>