Change in philosophy. ValidatorContext remains the main entry point for validation and ValidatorIndex.validateFor methods are retained for backward compatibility.  But ValidatorContext no longer publicly exposes Validator objects via a validateFor method.
ValidatorContext supports a few different styles of validation:
 - 	isValid(Object obj, ValidatorInterceptor i) supports no exception
	style validation.
 -	assertValid(Object obj, ValidatorInterceptor i) always throws an
	exception when validation failures occur (even if these were
	trapped by a ValidationInterceptor)
 - 	validate(Object obj, ValidatorInterceptor i) may or may not throw
 	an exception depending on the interceptor used.

ValidatorContext hides the ValidatorIndex (and by implication the
Validator instances it manages) in a ValidatorExecutionContext that is
made available to generated sources, but isn't visible to API
consumers.

Backward compatibility is maintained in many cases, but it was unavoidable
to change the ValidatorIndex API.  Implementers of that API will need
to rework their code.

Signed-off-by: Ross Lamont <rlamont@componentcorp.com>