fix golang oneof with a typed-nil value causes panic.
The existing generated code for a oneof field causes a nil-panic if
the field contains a typed-nil value. Go correctly infers the value
is one of the switch cases but the validation code assumes the
value is not a typed-nil.

THERE IS AN ARGUMENT THAT THIS BEHAVIOUR IS CORRECT. The authors of
go protobuf have stated that any typed-nil in a oneof field is
invalid, and so this maybe should cause a panic. Nil panic however
is not a helpful error message, and maybe the behaviour should be
given as a option to the invoker.