Improve Java UUID Validation Performance (#670)
Replaces the existing UUID validation implementation with the UUID
implementation that was added to Java in version 15. This implementation
is significantly faster than our previous implementation, and requires
fewer allocations (in fact only one global allocation, which is lazy and
cached from there on out). Only SIMD vector instructions could be used
to make it even faster.

Co-authored-by: Elliot Jackson <13633636+ElliotMJackson@users.noreply.github.com>